requirements.txt

django==2.2.3
django-crispy-forms==1.7.2
django-bootstrap4==0.0.8
six==1.10.0